+++ src/elements-experimental/seat-reservation/common/index.ts
@@ -1,3 +1,3 @@
 export * from './mapper.js';
 export * from './navigation-icon-mapper.js';
-export * from './i18n/i18n.js';
+export * from './i18n.js';
