+++ src/elements-experimental/seat-reservation/seat-reservation-place-control/seat-reservation-place-control.ts
@@ -17,10 +17,10 @@ import '../seat-reservation-graphic.js';
 import style from './seat-reservation-place-control.scss?lit&inline';
 
 /**
- * Describe the purpose of the component with a single short sentence.
+ * Output the graphic of a seat or a bicycle place as a control element.
  *
- * @slot - Use the unnamed slot to add elements.
- * @event {CustomEvent<PlaceSelection>} selectPlace - Emits when select a place and returns a PlaceSelection object with necessary place information
+ * event {CustomEvent<PlaceSelection>} selectPlace - Emits when select a place and returns a PlaceSelection object with necessary place information
+ * @event {CustomEvent<void>} selectPlace - Emits when an place was selected by user.
  */
 export
 @customElement('sbb-seat-reservation-place-control')
@@ -32,12 +32,12 @@ class SbbSeatReservationPlaceControlElement extends LitElement {
 
   /** Type Prop */
   @forceType()
-  @property({ attribute: 'type' })
+  @property({ attribute: 'type', type: String })
   public accessor type: PlaceType = 'SEAT';
 
   /** State Prop */
   @forceType()
-  @property({ attribute: 'state' })
+  @property({ attribute: 'state', type: String })
   public accessor state: PlaceState = 'FREE';
 
   /** Rotation Prop */
@@ -133,7 +133,7 @@ class SbbSeatReservationPlaceControlElement extends LitElement {
 
     return html`
       <div
-        class="sbb-seat-reservation-place-control sbb-seat-reservation-place-control--type-${type}  sbb-seat-reservation-place-control--state-${state}"
+        class="sbb-seat-reservation-place-control sbb-seat-reservation-place-control--type-${type} sbb-seat-reservation-place-control--state-${state}"
       >
         <button
           class="sbb-seat-reservation-place-control__button"
